Move v1 secret functions in module_utils
