+++ ansible/plugins/module_utils/load_secrets_v1.py
